+++ app/commands/characters_context/change_feat_command.rb
@@ -12,6 +12,7 @@ module CharactersContext
 
       params do
         required(:character_feat).filled(type?: ::Character::Feat)
+        optional(:active).filled(:bool)
         optional(:used_count).filled(:integer)
         optional(:value)
       end