+++ app/decorators/daggerheart_character/features_decorator.rb
@@ -1,0 +1,86 @@
+# frozen_string_literal: true
+
+module DaggerheartCharacter
+  class FeaturesDecorator
+    attr_accessor :wrapped
+
+    def initialize(obj)
+      @wrapped = obj
+    end
+
+    def method_missing(method, *_args)
+      if instance_variable_defined?(:"@#{method}")
+        instance_variable_get(:"@#{method}")
+      else
+        instance_variable_set(:"@#{method}", wrapped.public_send(method))
+      end
+    end
+
+    # rubocop: disable Metrics/AbcSize, Metrics/MethodLength, Metrics/PerceivedComplexity
+    def features
+      @features ||=
+        available_features.filter_map do |feature|
+          # добавлять статические бонусы или включенные
+          if feature_bonuses_enabled?(feature)
+            feature.feat.eval_variables.each do |method_name, variable|
+              result = eval_variable(feature.feat, variable)
+              instance_variable_set(:"@#{method_name}", result) if result
+            end
+          end
+          next if feature.feat.kind == 'update_result'
+
+          feature.feat.description_eval_variables.transform_values! do |value|
+            eval_variable(feature.feat, value) || value
+          end
+          {
+            id: feature.id,
+            slug: feature.feat.slug,
+            kind: feature.feat.kind,
+            title: feature.feat.title[I18n.locale.to_s],
+            description: update_feature_description(feature),
+            limit: feature.feat.description_eval_variables['limit'],
+            used_count: feature.used_count,
+            limit_refresh: feature.feat.limit_refresh,
+            options: feature.feat.options,
+            value: feature.value,
+            origin: feature.feat.origin,
+            active: feature.active,
+            continious: feature.feat.continious
+          }.compact
+        end
+    end
+    # rubocop: enable Metrics/AbcSize, Metrics/MethodLength, Metrics/PerceivedComplexity
+
+    private
+
+    def feature_bonuses_enabled?(feature)
+      ready_to_use = feature.value.is_a?(Hash) ? feature.value.fetch('ready_to_use', true) : true
+      (!feature.feat.continious && ready_to_use) || feature.active
+    end
+
+    def update_feature_description(feature)
+      result = feature.feat.description[I18n.locale.to_s]
+      feature.feat.description_eval_variables.each { |key, value| result.gsub!("{{#{key}}}", value.to_s) }
+      result
+    end
+
+    # rubocop: disable Security/Eval, Style/MethodCalledOnDoEndBlock
+    def eval_variable(feat, variable)
+      lambda do
+        eval(variable)
+      end.call
+    rescue StandardError, SyntaxError => e
+      monitoring_feat_error(e, feat)
+      nil
+    end
+    # rubocop: enable Security/Eval, Style/MethodCalledOnDoEndBlock
+
+    def monitoring_feat_error(exception, feat)
+      Charkeeper::Container.resolve('monitoring.client').notify(
+        exception: Monitoring::FeatVariableError.new('Feat variable error'),
+        metadata: { slug: feat.slug, message: exception.message },
+        severity: :info
+      )
+    end
+  end
+end