+++ app/serializers/item_serializer.rb
@@ -3,7 +3,7 @@
 class ItemSerializer < ApplicationSerializer
   ATTRIBUTES = %i[id slug kind name data info].freeze
 
-  attributes :id, :slug, :kind, :name, :data, :info
+  attributes(*ATTRIBUTES)
 
   def name
     object.name[I18n.locale.to_s]