+++ db/migrate/20251016065220_add_bonus_eval_variables_to_feats.rb
@@ -1,0 +1,5 @@
+class AddBonusEvalVariablesToFeats < ActiveRecord::Migration[8.0]
+  def change
+    add_column :feats, :bonus_eval_variables, :jsonb
+  end
+end